+++ tests/test-text-hover.php
@@ -188,7 +188,7 @@ class Text_Hover_Test extends WP_UnitTestCase {
 	}
 
 	public function test_version() {
-		$this->assertEquals( '3.9', c2c_TextHover::get_instance()->version() );
+		$this->assertEquals( '3.9.1', c2c_TextHover::get_instance()->version() );
 	}
 
 	public function test_instance_object_is_returned() {