+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Unit tests:
     * Change: Implement a more generic approach to capture default values provided for a filter
     * New: Add test to verify the lack of any defined hover text doesn't remove zeroes from text
+    * Fix: Correct typo in function name used
 
 ## 3.9 _(2020-01-08)_